sharder/requirements.txt
oauthlib
cryptography
psycopg2-binary
pycurl
pymysql
pyyaml
ruamel.yaml
sqlalchemy
tornado
